requirements.txt
SQLAlchemy~=1.4.41
PyMySQL~=1.0.2
cryptography~=38.0.1
python-dotenv~=0.21.0